몬스터 개발 2
- 데미지 및 사망 처리 구현
- 소화기 자폭 스킬 추가
- 거리 조절 이동 수정
- 기타 자료구조 수정

+++ Assets/Scripts/Monster/MobScripts/MobObj.cs
@@ -30,8 +30,8 @@ public class MobObj : ScriptableObject
 
 
     [Header("States")]
-    public float minHealth;   // 최소 체력
-    public float maxHealth;   // 최대 체력
+    public int minHealth;   // 최소 체력
+    public int maxHealth;   // 최대 체력
 
     public int attackDamage; // 공격력
